+++ src/ResultProcessor.php
@@ -4,10 +4,15 @@ declare(strict_types=1);
 
 namespace Magdicom;
 
+/**
+ * @template TRaw of mixed
+ * @template-covariant TProcessed of mixed
+ */
 interface ResultProcessor
 {
     /**
-     * @param list<mixed> $results
+     * @param list<TRaw> $results
+     * @return TProcessed
      */
     public function process(array $results, ProcessingContext $context): mixed;
 }